requirements.txt
boto3==1.38.42
boto3-stubs==1.38.42
bottle==0.13.4
certifi==2025.6.15
cryptography==45.0.4
gunicorn==23.0.0
mypy-boto3-s3==1.38.39
node-semver==0.9.0
PyJWT==2.10.1
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
types-boto3-s3==1.38.39
urllib3==2.5.0
